Introdcuing `towncrier` to MCEq development (#109)
towncrier is a small tool capable of generating a changelog with ease.
For each PR the author of the PR has to create a so called "newsfragment". This fragment should contain a meaningful entry for the changelog on what this PR is doing!
Newsfragments are store in the new changes directory.